moved configs to test/configs
