+++ test/configs/FileResourceStoreConfig.ts
@@ -13,25 +13,25 @@ import {
   ResourceStore,
   ResponseDescription,
   RuntimeConfig,
-  ServerConfig,
   TurtleToQuadConverter,
-} from '..';
-import { UnsecureWebIdExtractor } from '../src/authentication/UnsecureWebIdExtractor';
-import { AllowEverythingAuthorizer } from '../src/authorization/AllowEverythingAuthorizer';
-import { UrlBasedAclManager } from '../src/authorization/UrlBasedAclManager';
-import { BasicRequestParser } from '../src/ldp/http/BasicRequestParser';
-import { BasicResponseWriter } from '../src/ldp/http/BasicResponseWriter';
-import { BasicTargetExtractor } from '../src/ldp/http/BasicTargetExtractor';
-import { RawBodyParser } from '../src/ldp/http/RawBodyParser';
-import { DeleteOperationHandler } from '../src/ldp/operations/DeleteOperationHandler';
-import { GetOperationHandler } from '../src/ldp/operations/GetOperationHandler';
-import { PostOperationHandler } from '../src/ldp/operations/PostOperationHandler';
-import { PutOperationHandler } from '../src/ldp/operations/PutOperationHandler';
-import { FileResourceStore } from '../src/storage/FileResourceStore';
+} from '../..';
+import { UnsecureWebIdExtractor } from '../../src/authentication/UnsecureWebIdExtractor';
+import { AllowEverythingAuthorizer } from '../../src/authorization/AllowEverythingAuthorizer';
+import { UrlBasedAclManager } from '../../src/authorization/UrlBasedAclManager';
+import { BasicRequestParser } from '../../src/ldp/http/BasicRequestParser';
+import { BasicResponseWriter } from '../../src/ldp/http/BasicResponseWriter';
+import { BasicTargetExtractor } from '../../src/ldp/http/BasicTargetExtractor';
+import { RawBodyParser } from '../../src/ldp/http/RawBodyParser';
+import { DeleteOperationHandler } from '../../src/ldp/operations/DeleteOperationHandler';
+import { GetOperationHandler } from '../../src/ldp/operations/GetOperationHandler';
+import { PostOperationHandler } from '../../src/ldp/operations/PostOperationHandler';
+import { PutOperationHandler } from '../../src/ldp/operations/PutOperationHandler';
+import { FileResourceStore } from '../../src/storage/FileResourceStore';
+import { ServerConfig } from '../configs/ServerConfig';
 
 // This is the configuration from bin/server.ts
 
-export class LockingResourceStore implements ServerConfig {
+export class FileResourceStoreConfig implements ServerConfig {
   public store: ResourceStore;
   public aclManager: AclManager;
