+++ test/integration/AuthenticatedLdpHandler.test.ts
@@ -2,8 +2,8 @@ import * as url from 'url';
 import { namedNode, quad } from '@rdfjs/data-model';
 import { Parser } from 'n3';
 import { MockResponse } from 'node-mocks-http';
-import { SimpleHandlersTestConfig } from '../../configs/SimpleHandlersTestConfig';
-import { SimpleTestConfig } from '../../configs/SimpleTestConfig';
+import { SimpleHandlersTestConfig } from '../configs/SimpleHandlersTestConfig';
+import { SimpleTestConfig } from '../configs/SimpleTestConfig';
 import { call } from '../util/Util';
 
 describe('An integrated AuthenticatedLdpHandler', (): void => {
